Update jsdom, add new dependencies and implement crawlPage function
This commit includes an update of the jsdom to version 23.2.0, removes the "main" attribute from package.json, and adds a new function named crawlPage in crawler.js module. The crawl.test.js file is updated to include semicolons at the end of the statements. The Node version in .nvmrc file is updated to 20.10.0. Also, the main.js file is updated to import and use the crawlPage function.